Reserve exactly the field, and let a centred display recentre
The room reserved for a field is now its measured box, growing and
shrinking with it, and a centred display recentres around it with every
keystroke. Keeping freed room until commit, and pinning the display in
place while editing, were each there to spare the reader movement; with
the expression drawn in step with the keystroke, the movement they
spared is smooth, and what they left behind — room that did not close
up on delete, a display that jumped to centre on Enter — was the odd
part. The slot's `commit()` goes with them: Enter has nothing to do for
the layout.
